# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Api to create a language ll1",
        "title": "GO-Api API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/compare/simulate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Simulate compiler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compare"
                ],
                "summary": "Simulate compiler with public resources",
                "parameters": [
                    {
                        "description": "Simulate object",
                        "name": "simulate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/compare.Simulate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/compare.SimResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/compare.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User object that needs to be created",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/login.GoodResponse"
                        }
                    },
                    "400": {
                        "description": "Error creating user",
                        "schema": {
                            "$ref": "#/definitions/login.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "Create a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "User object that needs to be created",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/login.GoodResponse"
                        }
                    },
                    "400": {
                        "description": "Error creating user",
                        "schema": {
                            "$ref": "#/definitions/login.BadResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/priv/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a Private scanner from a Yalex",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Create a Private scanner from a Yalex",
                "parameters": [
                    {
                        "description": "Yalex object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.YalexPublic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/priv/simulate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Simulate a scanner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Simulate a scanner",
                "parameters": [
                    {
                        "description": "Scanner object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.SimScan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/public/create": {
            "post": {
                "description": "Create a Public scanner from a Yalex",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Create a Public scanner from a Yalex",
                "parameters": [
                    {
                        "description": "Yalex object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.YalexPublic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scanners/public/simulate": {
            "post": {
                "description": "Simulate a scanner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalex"
                ],
                "summary": "Simulate a scanner",
                "parameters": [
                    {
                        "description": "Scanner object",
                        "name": "yalex",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scanners.ScannerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/scanners.SimScan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/yapar/priv/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a Private table SRL from a Yalp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalp"
                ],
                "summary": "Create a Private table SRL from a Yalp",
                "parameters": [
                    {
                        "description": "Yalp object",
                        "name": "yalp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.Yalp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/yapar/priv/get": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a table SRL from user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalp"
                ],
                "summary": "Get a table SRL from user",
                "parameters": [
                    {
                        "description": "Table object",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/yapar/pub/create": {
            "post": {
                "description": "Create a Public table SRL from a Yalp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalp"
                ],
                "summary": "Create a Public table SRL from a Yalp",
                "parameters": [
                    {
                        "description": "Yalp object",
                        "name": "yalp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.YalpPublic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/yapar/pub/get": {
            "post": {
                "description": "Get a table SRL from public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yalp"
                ],
                "summary": "Get a table SRL from public",
                "parameters": [
                    {
                        "description": "Table object",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.TableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/yaparroutes.BadRequestResponse"
                        }
                    }
                }
            }
        },
        "/image/{fileID}": {
            "get": {
                "description": "Get an image from GridFS by file ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "image"
                ],
                "summary": "Get an image from GridFS by file ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "fileID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/scanners.BadRequestResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "afd.ResponseSim": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "simulatedParts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/afd.SimulatedPart"
                    }
                },
                "stringSummary": {
                    "type": "string"
                }
            }
        },
        "afd.SimulatedPart": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "final": {
                    "type": "integer"
                },
                "init": {
                    "type": "integer"
                },
                "runes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "compare.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "compare.SimResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "scannerResult": {
                    "type": "string"
                },
                "sim": {
                    "$ref": "#/definitions/yapar.SimResponse"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "compare.Simulate": {
            "type": "object",
            "properties": {
                "contentSimulate": {
                    "type": "string"
                },
                "scannerName": {
                    "type": "string"
                },
                "slrName": {
                    "type": "string"
                }
            }
        },
        "login.BadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "login.GoodResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "login.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "1234"
                },
                "username": {
                    "type": "string",
                    "example": "daniel"
                }
            }
        },
        "scanners.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "scanners.ScannerRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "scannerName": {
                    "type": "string"
                }
            }
        },
        "scanners.ScannerResponse": {
            "type": "object",
            "properties": {
                "filesId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Scanner created successfully"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "scanners.SimScan": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "simPart": {
                    "$ref": "#/definitions/afd.ResponseSim"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "scanners.YalexPublic": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "yapar.Action": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "yapar.SLR": {
            "type": "object",
            "properties": {
                "ignored": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "table": {
                    "$ref": "#/definitions/yapar.Table"
                }
            }
        },
        "yapar.SimResponse": {
            "type": "object",
            "properties": {
                "accept": {
                    "type": "boolean"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/yapar.Step"
                    }
                }
            }
        },
        "yapar.Step": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "integer"
                },
                "state": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {}
                    }
                }
            }
        },
        "yapar.Table": {
            "type": "object",
            "additionalProperties": {
                "type": "object",
                "additionalProperties": {
                    "$ref": "#/definitions/yapar.Action"
                }
            }
        },
        "yaparroutes.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "yaparroutes.TableRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "yaparroutes.TableResponse": {
            "type": "object",
            "properties": {
                "imageURL": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Table created successfully"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "table": {
                    "$ref": "#/definitions/yapar.SLR"
                }
            }
        },
        "yaparroutes.Yalp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "yaparroutes.YalpPublic": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}